#13 Refactor forms and form fields
